Extract Movie response class
